contrib/masmx64/iffas8664.c: fix rare 32 vs 64 alignment issue
unsigned long is only 32 bit on win64, while hold (and size_t) are
64 bit.

If the caller sometimes changes the alignment of the input data
(for example, while moving data in a fixed input buffer to make room
for more data), then there are rarish cases that need to preserve
more than 32 bits for this alignment loop to work correctly.
Especially if both input and output are regularly using Z_SYNC_FLUSH,
to support bi-directional streaming message protocols without
arbitrarily long delays.

+++ contrib/masmx64/inffas8664.c
@@ -144,7 +144,7 @@ unsigned start;         /* inflate()'s starting value for strm->avail_out */
 
     /* align in on 1/2 hold size boundary */
     while (((size_t)(void *)ar.in & (sizeof(ar.hold) / 2 - 1)) != 0) {
-        ar.hold += (unsigned long)*ar.in++ << ar.bits;
+        ar.hold += (size_t)*ar.in++ << ar.bits;
         ar.bits += 8;
     }
